refactor(opencode): extract readiness retry policy
