+++ scripts/ci/source-file-size-baseline.json
@@ -77,7 +77,7 @@
     "src/main/services/team/provisioning/TeamProvisioningStreamEvents.ts": 1212,
     "src/main/services/team/ReviewApplierService.ts": 2062,
     "src/main/services/team/ReviewDecisionStore.ts": 1970,
-    "src/main/services/team/runtime/OpenCodeTeamRuntimeAdapter.ts": 1648,
+    "src/main/services/team/runtime/OpenCodeTeamRuntimeAdapter.ts": 1591,
     "src/main/services/team/TaskChangeComputer.ts": 1357,
     "src/main/services/team/TaskChangeLedgerReader.ts": 2159,
     "src/main/services/team/taskLogs/stream/BoardTaskLogStreamService.ts": 2516,
